Respond to Daniel's comments
